+++ x/gitopia/client/cli/tx.go
@@ -33,10 +33,14 @@ func GetTxCmd() *cobra.Command {
 
 	cmd.AddCommand(CmdCreatePullRequest())
 	cmd.AddCommand(CmdUpdatePullRequest())
+	cmd.AddCommand(CmdUpdatePullRequestTitle())
+	cmd.AddCommand(CmdUpdatePullRequestDescription())
+	cmd.AddCommand(CmdSetPullRequestState())
 	cmd.AddCommand(CmdDeletePullRequest())
 
 	cmd.AddCommand(CmdCreateOrganization())
 	cmd.AddCommand(CmdUpdateOrganizationMember())
+	cmd.AddCommand(CmdRemoveOrganizationMember())
 	cmd.AddCommand(CmdUpdateOrganization())
 	cmd.AddCommand(CmdDeleteOrganization())
 
@@ -55,6 +59,8 @@ func GetTxCmd() *cobra.Command {
 	cmd.AddCommand(CmdForkRepository())
 	cmd.AddCommand(CmdRenameRepository())
 	cmd.AddCommand(CmdChangeOwner())
+	cmd.AddCommand(CmdUpdateRepositoryCollaborator())
+	cmd.AddCommand(CmdRemoveRepositoryCollaborator())
 	cmd.AddCommand(CmdCreateBranch())
 	cmd.AddCommand(CmdSetDefaultBranch())
 	cmd.AddCommand(CmdDeleteBranch())