+++ x/gitopia/keeper/msg_server_pullRequest.go
@@ -2,7 +2,9 @@ package keeper
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"strconv"
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -12,6 +14,11 @@ import (
 
 func (k msgServer) CreatePullRequest(goCtx context.Context, msg *types.MsgCreatePullRequest) (*types.MsgCreatePullRequestResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Checks that the element exists
+	if !k.HasUser(ctx, msg.Creator) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("User %v doesn't exist", msg.Creator))
+	}
 
 	if !k.HasRepository(ctx, msg.HeadRepoId) {
 		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("headRepositoryId %d doesn't exist", msg.HeadRepoId))
@@ -103,6 +110,142 @@ func (k msgServer) UpdatePullRequest(goCtx context.Context, msg *types.MsgUpdate
 	k.SetPullRequest(ctx, pullRequest)
 
 	return &types.MsgUpdatePullRequestResponse{}, nil
+}
+
+func (k msgServer) UpdatePullRequestTitle(goCtx context.Context, msg *types.MsgUpdatePullRequestTitle) (*types.MsgUpdatePullRequestTitleResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Checks that the element exists
+	if !k.HasUser(ctx, msg.Creator) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("User %v doesn't exist", msg.Creator))
+	}
+
+	// Checks that the element exists
+	if !k.HasPullRequest(ctx, msg.Id) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("key %d doesn't exist", msg.Id))
+	}
+
+	// Checks if the the msg sender is the same as the current owner
+	if msg.Creator != k.GetPullRequestOwner(ctx, msg.Id) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner")
+	}
+
+	var pullrequest = k.GetPullRequest(ctx, msg.Id)
+
+	pullrequest.Title = msg.Title
+	pullrequest.UpdatedAt = time.Now().Unix()
+
+	k.SetPullRequest(ctx, pullrequest)
+
+	return &types.MsgUpdatePullRequestTitleResponse{}, nil
+}
+
+func (k msgServer) UpdatePullRequestDescription(goCtx context.Context, msg *types.MsgUpdatePullRequestDescription) (*types.MsgUpdatePullRequestDescriptionResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Checks that the element exists
+	if !k.HasUser(ctx, msg.Creator) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("User %v doesn't exist", msg.Creator))
+	}
+
+	// Checks that the element exists
+	if !k.HasPullRequest(ctx, msg.Id) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("key %d doesn't exist", msg.Id))
+	}
+
+	// Checks if the the msg sender is the same as the current owner
+	if msg.Creator != k.GetPullRequestOwner(ctx, msg.Id) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner")
+	}
+
+	var pullrequest = k.GetPullRequest(ctx, msg.Id)
+
+	pullrequest.Description = msg.Description
+	pullrequest.UpdatedAt = time.Now().Unix()
+
+	k.SetPullRequest(ctx, pullrequest)
+
+	return &types.MsgUpdatePullRequestDescriptionResponse{}, nil
+}
+
+func (k msgServer) SetPullRequestState(goCtx context.Context, msg *types.MsgSetPullRequestState) (*types.MsgSetPullRequestStateResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Checks that the element exists
+	if !k.HasUser(ctx, msg.Creator) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("User %v doesn't exist", msg.Creator))
+	}
+
+	// Checks that the element exists
+	if !k.HasPullRequest(ctx, msg.Id) {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("key %d doesn't exist", msg.Id))
+	}
+
+	var pullRequest = k.GetPullRequest(ctx, msg.Id)
+	currentTime := time.Now().Unix()
+
+	repository := k.GetRepository(ctx, pullRequest.BaseRepoId)
+
+	var o Owner
+	if err := json.Unmarshal([]byte(repository.Owner), &o); err != nil {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "unable to unmarshal owner")
+	}
+
+	if o.Type == "User" {
+		if !((msg.State == "Open" || msg.State == "Closed") && msg.Creator == pullRequest.Creator) &&
+			msg.Creator != o.ID && repository.Collaborators[msg.Creator] != "Admin" {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, fmt.Sprintf("user %v doesn't have permission to perform this operation", msg.Creator))
+		}
+	} else if o.Type == "Organization" {
+		orgId, err := strconv.ParseUint(o.ID, 10, 64)
+		if err != nil {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "can't fetch baseRepository owner")
+		}
+		if !k.HasOrganization(ctx, orgId) {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, fmt.Sprintf("organization %v doesn't exist", o.ID))
+		}
+
+		organization := k.GetOrganization(ctx, orgId)
+
+		// Checks if the the msg sender is the same as the current owner
+		if !((msg.State == "Open" || msg.State == "Closed") && msg.Creator == pullRequest.Creator) &&
+			organization.Members[msg.Creator] != "Owner" && repository.Collaborators[msg.Creator] != "Admin" {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, fmt.Sprintf("user %v doesn't have permission to perform this operation", msg.Creator))
+		}
+	} else {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("can't fetch baseRepository owner"))
+	}
+
+	switch msg.State {
+	case "Open":
+		if pullRequest.State == "Open" || pullRequest.State == "Merged" {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("can't open (%v) pullRequest", pullRequest.State))
+		}
+	case "Closed":
+		if pullRequest.State == "Closed" || pullRequest.State == "Merged" {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("can't close (%v) pullRequest", pullRequest.State))
+		}
+		pullRequest.ClosedAt = currentTime
+		pullRequest.ClosedBy = msg.Creator
+	case "Merged":
+		if pullRequest.State == "Merged" || pullRequest.State == "Closed" {
+			return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("can't merge (%v) pullRequest", pullRequest.State))
+		}
+		pullRequest.MergedAt = currentTime
+		pullRequest.MergedBy = msg.Creator
+		pullRequest.MergeCommitSha = msg.MergeCommitSha
+	default:
+		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("invalid state (%v)", msg.State))
+	}
+
+	pullRequest.State = msg.State
+	pullRequest.UpdatedAt = currentTime
+
+	k.SetPullRequest(ctx, pullRequest)
+
+	return &types.MsgSetPullRequestStateResponse{
+		State: pullRequest.State,
+	}, nil
 }
 
 func (k msgServer) DeletePullRequest(goCtx context.Context, msg *types.MsgDeletePullRequest) (*types.MsgDeletePullRequestResponse, error) {