+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes will be documented here.
 
+## [Unreleased] - yyyy-mm-dd
+
+### Added
+
+- Added transaction to set pullRequest state
+- Added transaction to update pullRequest description
+- Added transaction to update pullRequest title
+- Added transaction to Remove organization member
+- Added transaction to Remove repository collaborator
+- Added transaction to Add/Update repository collaborator
+- Added api to get all user organizations (no pagination)
+- Added api to get organization by name
+- Added api to query Organization all repositories (with pagination)
+- Added api to get organization repository
+- Added api to get sha by branchName
+- Implemented transaction to change ownership of a repository
+
+### Changed
+
+- Removed field `repositoryNames` and `repositories_archived` from User and Organization
+- Fields `repositories` and `organizations` are changed to map
+
 ## [0.6.0] - 2021-08-11
 
 ### Added